Data-Pipelining/streamlit-dashboard/requirements.txt
streamlit
Pillow
matplotlib
